# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "The user has been successfully created." },
          "400": { "description": "Bad Request." },
          "403": { "description": "Forbidden." },
          "409": { "description": "Conflict - Email already exists." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create a new user",
        "tags": ["users"]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "firstName",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "lastName",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "organizationId",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "Return all users." } },
        "security": [{ "bearer": [] }],
        "summary": "Get all users",
        "tags": ["users"]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Return the user." },
          "404": { "description": "User not found." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get user by ID",
        "tags": ["users"]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "The user has been successfully updated." },
          "400": { "description": "Bad Request." },
          "403": { "description": "Forbidden." },
          "404": { "description": "User not found." },
          "409": { "description": "Conflict - Email already exists." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update user by ID",
        "tags": ["users"]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "The user has been successfully deleted." },
          "400": { "description": "Bad Request." },
          "403": { "description": "Forbidden." },
          "404": { "description": "User not found." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Delete user by ID",
        "tags": ["users"]
      }
    },
    "/users/{id}/roles": {
      "post": {
        "operationId": "UsersController_assignRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AssignRoleDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "The role has been successfully assigned." },
          "400": { "description": "Bad Request." },
          "403": { "description": "Forbidden." },
          "404": { "description": "User or Role not found." },
          "409": { "description": "Conflict - User already has this role." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Assign a role to a user",
        "tags": ["users"]
      }
    },
    "/users/{id}/roles/{roleId}": {
      "delete": {
        "operationId": "UsersController_removeRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "The role has been successfully removed." },
          "400": { "description": "Bad Request." },
          "403": { "description": "Forbidden." },
          "404": { "description": "User or Role not found." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Remove a role from a user",
        "tags": ["users"]
      }
    },
    "/users/{id}/departments": {
      "post": {
        "operationId": "UsersController_updateDepartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The department has been successfully assigned or updated."
          },
          "400": { "description": "Bad Request." },
          "403": { "description": "Forbidden." },
          "404": { "description": "User or Department not found." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Assign or update a department for a user",
        "tags": ["users"]
      }
    },
    "/users/{id}/departments/{departmentId}": {
      "delete": {
        "operationId": "UsersController_removeDepartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "departmentId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "The department has been successfully removed."
          },
          "400": { "description": "Bad Request." },
          "403": { "description": "Forbidden." },
          "404": { "description": "User or Department not found." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Remove a department from a user",
        "tags": ["users"]
      }
    },
    "/users/{id}/permissions": {
      "get": {
        "operationId": "UsersController_getUserPermissions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Return user permissions." },
          "404": { "description": "User not found." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get user permissions",
        "tags": ["users"]
      },
      "post": {
        "operationId": "UsersController_assignPermission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignUserPermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The permission has been successfully assigned."
          },
          "403": { "description": "Forbidden." },
          "404": { "description": "User or Permission not found." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Assign a permission to a user",
        "tags": ["users"]
      }
    },
    "/users/{id}/permissions/{permissionId}": {
      "delete": {
        "operationId": "UsersController_removePermission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "permissionId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "The permission has been successfully removed."
          },
          "403": { "description": "Forbidden." },
          "404": { "description": "User or Permission not found." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Remove a permission from a user",
        "tags": ["users"]
      }
    },
...
  },
  "info": {
    "title": "Multi-Sector Management API",
    "description": "API for managing organizations across multiple sectors",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "UserDepartmentDto": {
        "type": "object",
        "properties": {
          "departmentId": { "type": "string", "description": "Department ID" },
          "position": {
            "type": "string",
            "description": "User position in department"
          },
          "isHead": {
            "type": "boolean",
            "description": "Whether user is head of department"
          }
        },
        "required": ["departmentId"]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "description": "User email" },
          "password": { "type": "string", "description": "User password" },
          "firstName": { "type": "string", "description": "User first name" },
          "lastName": { "type": "string", "description": "User last name" },
          "isActive": {
            "type": "boolean",
            "description": "User active status",
            "default": true
          },
          "organizationId": {
            "type": "string",
            "description": "Organization ID"
          },
          "roles": {
            "description": "Role IDs to assign",
            "type": "array",
            "items": { "type": "string" }
          },
          "departments": {
            "description": "Departments to assign",
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserDepartmentDto" }
          },
          "sectorProfile": {
            "type": "object",
            "description": "Sector-specific user profile data"
          },
          "address": { "type": "object", "description": "User address" }
        },
        "required": ["email", "password", "firstName", "lastName"]
      },
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "description": "User email" },
          "password": { "type": "string", "description": "User password" },
          "firstName": { "type": "string", "description": "User first name" },
          "lastName": { "type": "string", "description": "User last name" },
          "isActive": {
            "type": "boolean",
            "description": "User active status",
            "default": true
          },
          "organizationId": {
            "type": "string",
            "description": "Organization ID"
          },
          "roles": {
            "description": "Role IDs to assign",
            "type": "array",
            "items": { "type": "string" }
          },
          "departments": {
            "description": "Departments to assign",
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserDepartmentDto" }
          },
          "sectorProfile": {
            "type": "object",
            "description": "Sector-specific user profile data"
          },
          "address": { "type": "object", "description": "User address" }
        }
      },
      "AssignRoleDto": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string",
            "description": "Role ID to assign to the user"
          },
          "validUntil": {
            "format": "date-time",
            "type": "string",
            "description": "Expiration date for the role (temporary role)"
          }
        },
        "required": ["roleId"]
      },
      "UpdateUserDepartmentDto": {
        "type": "object",
        "properties": {
          "departmentId": { "type": "string", "description": "Department ID" },
          "position": {
            "type": "string",
            "description": "User position in department"
          },
          "isHead": {
            "type": "boolean",
            "description": "Whether user is head of department"
          }
        },
        "required": ["departmentId"]
      },
      "AssignUserPermissionDto": {
        "type": "object",
        "properties": {
          "permissionId": {
            "type": "string",
            "description": "Permission ID to assign to the user"
          },
          "conditions": {
            "type": "object",
            "description": "Conditions for the permission (CASL conditions)"
          },
          "validUntil": {
            "format": "date-time",
            "type": "string",
            "description": "Expiration date for the permission (temporary access)"
          }
        },
        "required": ["permissionId"]
      },
...
    }
  }
}